worker/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --no-cache pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via httpx
backoff==2.2.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
cosl==0.0.56
    # via tempo-worker-k8s-operator (pyproject.toml)
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
googleapis-common-protos==1.68.0
    # via opentelemetry-exporter-otlp-proto-http
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via lightkube
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==6.11.0
    # via opentelemetry-api
lightkube==0.17.1
    # via cosl
lightkube-models==1.32.0.8
    # via lightkube
opentelemetry-api==1.21.0
    # via
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
opentelemetry-exporter-otlp-proto-common==1.21.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.21.0
    # via tempo-worker-k8s-operator (pyproject.toml)
opentelemetry-proto==1.21.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.21.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.42b0
    # via opentelemetry-sdk
ops==2.18.1
    # via
    #   tempo-worker-k8s-operator (pyproject.toml)
    #   cosl
protobuf==4.25.6
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydantic==2.10.6
    # via
    #   tempo-worker-k8s-operator (pyproject.toml)
    #   cosl
pydantic-core==2.27.2
    # via pydantic
pyyaml==6.0.2
    # via
    #   cosl
    #   lightkube
    #   ops
requests==2.32.3
    # via opentelemetry-exporter-otlp-proto-http
sniffio==1.3.1
    # via anyio
tenacity==9.0.0
    # via cosl
typing-extensions==4.12.2
    # via
    #   anyio
    #   cosl
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
websocket-client==1.8.0
    # via ops
wrapt==1.17.2
    # via deprecated
zipp==3.21.0
    # via importlib-metadata
